removed some warnings
git-svn-id: https://ffdshow-tryout.svn.sourceforge.net/svnroot/ffdshow-tryout/trunk@1583 3b938f2f-1a1a-0410-8054-a526ea5ff92c

+++ src/subtitles/TsubtitleProps.cpp
@@ -108,8 +108,8 @@ void TSubtitleProps::fix_size(LOGFONT &lf, HDC hdc, TfontManager *fontManager) c
   }
  else
   {
-   m_ascent64=lf.lfHeight*2*0.8;
-   m_descent64=lf.lfHeight*2-m_ascent64;
+   m_ascent64=int(lf.lfHeight*2*0.8);
+   m_descent64=int(lf.lfHeight*2-m_ascent64);
   }
 }
